+++ tests/test_codegen/testcases.py
@@ -576,7 +576,6 @@ NodeInterfaceFieldTestCase = QtGqlTestCase(
     test_name="NodeInterfaceFieldTestCase",
 )
 
-
 NonNodeUnionTestCase = QtGqlTestCase(
     schema=schemas.non_node_union.schema,
     operations="""
@@ -698,7 +697,6 @@ ListOfNonNodeType = QtGqlTestCase(
     test_name="ListOfNonNodeType",
 )
 
-
 ListOfUnionTestCase = QtGqlTestCase(
     schema=schemas.list_of_union.schema,
     operations="""
@@ -777,7 +775,6 @@ FragmentTestCase = QtGqlTestCase(
     """,
     test_name="FragmentTestCase",
 )
-
 
 FragmentsOnInterfaceTestCase = QtGqlTestCase(
     schema=schemas.non_node_interface_field.schema,
@@ -1027,6 +1024,32 @@ ListOfInputObjectTestCase = QtGqlTestCase(
     ),
 )
 
+InterfaceWithObjectField = QtGqlTestCase(
+    schema=schemas.interface_with_object_field.schema,
+    operations="""
+    query AnimalQuery($kind: AnimalKind!) {
+      animal(kind: $kind) {
+        metadata{
+          kind
+            gender
+            age
+        }
+        id
+        ... on Person {
+          language
+        }
+        ... on Dog {
+          furColor
+        }
+      }
+    }    """,
+    test_name="InterfaceWithObjectField",
+    metadata=TestCaseMetadata(
+        should_test_updates=BoolWithReason.false(
+            "There is no reason it would fail, this only failed due to object types being generated after interfaces.",
+        ),
+    ),
+)
 
 all_test_cases = [
     ScalarsTestCase,
@@ -1064,6 +1087,7 @@ all_test_cases = [
     ListOfInputObjectTestCase,
     OptionalInputTestCase,
     RecursiveInputObjectTestCase,
+    InterfaceWithObjectField,
     CustomUserScalarTestCase,
     ObjectsThatReferenceEachOtherTestCase,
     RootListOfTestCase,
@@ -1105,6 +1129,7 @@ implemented_testcases = [
     ListOfInputObjectTestCase,
     OptionalInputTestCase,
     RecursiveInputObjectTestCase,
+    InterfaceWithObjectField,
 ]